refactor(ExamLab): 移除文件路径显示转换器
- 从所有XAML文件中移除FilePathDisplayConverter的使用
- 恢复TextBox绑定为原始的直接绑定（不使用Converter）
- 删除FilePathDisplayConverter.cs文件
- 原因：输入框应该只接受和显示相对路径，不需要转换

修改文件：
- ExamLab/App.xaml - 移除转换器注册
- ExamLab/Views/ExcelModuleView.xaml - 恢复原始绑定
- ExamLab/Views/WordModuleView.xaml - 恢复原始绑定
- ExamLab/Views/PowerPointModuleView.xaml - 恢复原始绑定
- ExamLab/Views/CSharpModuleView.xaml - 恢复原始绑定
- ExamLab/Views/OperationPointConfigView.xaml - 恢复原始绑定

+++ ExamLab/Models/ImportExport/SpecializedExamExportDto.cs
@@ -85,7 +85,7 @@ public class SpecializedExamDto
     /// 总分
     /// </summary>
     [JsonPropertyName("totalScore")]
-    public int TotalScore { get; set; } = 100;
+    public double TotalScore { get; set; } = 100;
 
     /// <summary>
     /// 考试时长（分钟）
