+++ ExamLab/Models/OperationPoint.cs
@@ -306,7 +306,7 @@ public class OperationPoint : ReactiveObject
     /// <summary>
     /// 操作点分值
     /// </summary>
-    [Reactive] public decimal Score { get; set; }
+    [Reactive] public double Score { get; set; }
 
     /// <summary>
     /// 关联的评分题目ID
